+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/superset-core/GlobalFilter.tsx
