+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/superset-core/index.ts
@@ -1,0 +1,4 @@
+export * from './GlobalFilter';
+export * from './SelectPageSize';
+export * from './Pagination';
+export * from './SearchInput';
