+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/plugin/buildQuery.ts
@@ -12,7 +12,7 @@ import {
 import { PostProcessingRule } from '@superset-ui/core/src/query/types/PostProcessing';
 import { BuildQuery } from '@superset-ui/core/src/chart/registries/ChartBuildQueryRegistrySingleton';
 import { TableChartFormData } from '../superset-core-utils';
-import { updateExternalFormData } from '../superset-core-utils/externalAPIs';
+import { updateExternalFormData } from '../superset-core-utils/external-apis';
 
 /**
  * Infer query mode from form data. If `all_columns` is set, then raw records mode,
@@ -53,10 +53,8 @@ const buildQuery: BuildQuery<TableChartFormData> = (
   }
 
   formDataCopy.columns = formDataCopy.cols || [];
-  console.log('formDataCopy:', formDataCopy);
 
   return buildQueryContext(formDataCopy, baseQueryObject => {
-    console.log('baseQueryObject:', baseQueryObject);
     let { metrics, orderby = [], columns = [] } = baseQueryObject;
     let postProcessing: PostProcessingRule[] = [];
