+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/superset-core-utils/consts.ts
