+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/superset-core-utils/index.ts
@@ -3,3 +3,5 @@ export * from './format-column-value';
 export * from './get-shared-style';
 export * from './process-columns';
 export * from './use-async-state';
+export * from './consts';
+export * from './external-apis';
